Enhancing the Sending side of the socket to also minimize memory utilization and allocations.
Similiar to the Receive loop we created, the Send loop
will send N bytes and then call a callback method.

In this case, we queue up writes and dequeue in a loop. When there are
no more writes, we return. A RegisteredWaitHandle handles spanning our
thread to flush the queued writes to the underlying socket.

This allows us to have concurrent reading/writing.